gulpfile build-all
Changed the way that the gulpfile works, so that it now wipes .dist and
rebuilds on any file changes. That seems to make the most sense to me,
barring a more granular approach. Actually, the way to do it might be to
include a clean call (`del('dist/**/*.xyz`) in each of the build tasks.
But this works fine for my purposes.

Also fixed the problem of nodemon trying to find the server before it's
built. Server is the only task that nodemon waits on.