Removed client server
The server for the client is just a dumb http server, so I took out the
existing server code. Don't need any of that. In development i'm using
the npm http-server package. In production, could either switch to nginx
or continue using the http-server (which claims to be up to it, but idk.
it'd be worthwhile to learn nginx anyway).

Anyway, changed the file tree too, so there's no more 'public' /
'server' division. just public files.