Importing modules
Okay, so by just importing the various files, we bring them into the browserify stream. That means that we can have a mix of ES6 classes and ES5, which will make refactoring much easier. For now, the next thing to do is to pull in all of the services / controllers with import statements. Once it's all working, we can start flipping things over to classes, etc